+++ db-acf-ui.php
@@ -2,10 +2,10 @@
 /*
 Plugin Name: DB ACF Extension
 Description: Aangepaste ACF interface voor Digitale Bazen
-Version: 1.7.1
+Version: 1.7.2
 Author: Digitale Bazen
 Text Domain: db-acf-ui
-Update URI: bitbucket.org/digitale-bazen/db-acf-extension
+Update URI: https://github.com/yanickvdkorst/DB-ACF-Extension
 
 Icon: assets/icon-128x128.png
 Banner: assets/banner-772x250.png
@@ -21,7 +21,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  * Constants
  * ---------------------------
  */
-define( 'DB_ACF_UI_VERSION', '1.7.1' );
+define( 'DB_ACF_UI_VERSION', '1.7.2' );
 define( 'DB_ACF_UI_MIN_PHP_VERSION', '8.0' );
 
 define( 'DB_ACF_UI_FILE', __FILE__ );
@@ -72,90 +72,7 @@ add_action( 'plugins_loaded', function () {
  * ---------------------------
  * GitHub updater
  * ---------------------------
+ * Buiten de ACF-check gehouden: updates moeten ook werken als ACF (tijdelijk)
+ * niet actief is. Zie classes/updater.php.
  */
-
-// Inject update info
-add_filter( 'pre_set_site_transient_update_plugins', function ( $transient ) {
-
-    if ( empty( $transient->checked ) ) {
-        return $transient;
-    }
-
-    $plugin_slug = plugin_basename( DB_ACF_UI_FILE );
-
-    $response = wp_remote_get(
-        'https://api.github.com/repos/' . DB_ACF_UI_GITHUB_REPO . '/releases/latest',
-        [
-            'headers' => [ 'User-Agent' => 'WordPress' ],
-            'timeout' => 20,
-        ]
-    );
-
-    if ( is_wp_error( $response ) ) {
-        return $transient;
-    }
-
-    $data = json_decode( wp_remote_retrieve_body( $response ) );
-
-    if ( empty( $data->tag_name ) ) {
-        return $transient;
-    }
-
-    $remote_version = ltrim( $data->tag_name, 'v' );
-
-    if ( version_compare( $remote_version, DB_ACF_UI_VERSION, '>' ) ) {
-
-        $transient->response[ $plugin_slug ] = (object) [
-            'slug'        => DB_ACF_UI_SLUG,
-            'plugin'      => $plugin_slug,
-            'new_version' => $remote_version,
-            'url'         => $data->html_url,
-            'package' => sprintf(
-                'https://github.com/%s/releases/download/%s/db-acf-extension.zip',
-                DB_ACF_UI_GITHUB_REPO,
-                $data->tag_name
-            ),
-        ];
-    }
-
-    return $transient;
-} );
-
-// Plugin info popup
-add_filter( 'plugins_api', function ( $res, $action, $args ) {
-
-    if ( $action !== 'plugin_information' || $args->slug !== DB_ACF_UI_SLUG ) {
-        return $res;
-    }
-
-    $response = wp_remote_get(
-        'https://api.github.com/repos/' . DB_ACF_UI_GITHUB_REPO . '/releases/latest',
-        [
-            'headers' => [ 'User-Agent' => 'WordPress' ],
-            'timeout' => 20,
-        ]
-    );
-
-    if ( is_wp_error( $response ) ) {
-        return $res;
-    }
-
-    $data = json_decode( wp_remote_retrieve_body( $response ) );
-
-    return (object) [
-        'name'          => 'DB ACF Extension',
-        'slug'          => DB_ACF_UI_SLUG,
-        'version'       => ltrim( $data->tag_name, 'v' ),
-        'author'        => 'Digitale Bazen',
-        'homepage'      => $data->html_url,
-        'download_link' => sprintf(
-            'https://github.com/%s/releases/download/%s/db-acf-extension.zip',
-            DB_ACF_UI_GITHUB_REPO,
-            $data->tag_name
-        ),
-        'sections'      => [
-            'description' => 'Aangepaste ACF interface voor Digitale Bazen.',
-            'changelog'   => $data->body ?: 'Geen changelog opgegeven.',
-        ],
-    ];
-}, 20, 3 );
+\DB_ACF_UI\Updater::register();